requirements.txt
bcrypt
requests
gql
graphql-core
pandas
numpy
lxml==5.3.0
psycopg2
requests_toolbelt
soccerdata
streamlit
setuptools
sqlalchemy
thefuzz
